+++ pkg/server/server_test.go
@@ -861,7 +861,7 @@ func BenchmarkOpenFGAServer(b *testing.B) {
 				testDatastore := storagefixtures.RunDatastoreTestContainer(b, "postgres", imageVersion)
 
 				uri := testDatastore.GetConnectionURI(true)
-				ds, err := postgres.New(uri, sqlcommon.NewConfig())
+				ds, err := postgres.New(uri, sqlcommon.NewConfig(sqlcommon.WithMaxOpenConns(10)))
 				require.NoError(b, err)
 				b.Cleanup(ds.Close)
 				test.RunAllBenchmarks(b, ds)
@@ -879,7 +879,7 @@ func BenchmarkOpenFGAServer(b *testing.B) {
 		testDatastore := storagefixtures.RunDatastoreTestContainer(b, "mysql", "")
 
 		uri := testDatastore.GetConnectionURI(true)
-		ds, err := mysql.New(uri, sqlcommon.NewConfig())
+		ds, err := mysql.New(uri, sqlcommon.NewConfig(sqlcommon.WithMaxOpenConns(10)))
 		require.NoError(b, err)
 		b.Cleanup(ds.Close)
 		test.RunAllBenchmarks(b, ds)