+++ CHANGELOG.md
@@ -11,8 +11,21 @@ Try to keep listed changes to a concise bulleted list of simple explanations of 
 - test suite support for Postgres v17
 
 ### Fixed
+- Ensure `fanin.Stop` and `fanin.Drain` are called for all clients which may create blocking goroutines. [#2441](https://github.com/openfga/openfga/pull/2441)
+- Prevent throttled Go routines from "leaking" when a request context has been canceled or deadline exceeded. [#2450](https://github.com/openfga/openfga/pull/2450)
+
+## [1.8.12] - 2025-05-12
+[Full changelog](https://github.com/openfga/openfga/compare/v1.8.11...v1.8.12)
+
+### Changed
+- `DefaultResolveNodeBreadthLimit` changed from 100 to 10 in order to reduce connection contention. [#2425](https://github.com/openfga/openfga/pull/2425)
+- PostgreSQL and MySQL based iterators will load tuples only when needed (lazy loading). [#2425](https://github.com/openfga/openfga/pull/2425)
+
+### Fixed
 - Replace hardcoded Prometheus datasource UID (`PBFA97CFB590B2093`) with `${DS_PROMETHEUS}` in `telemetry/grafana/dashboards/openfga.json`. This allows the Grafana dashboard to correctly reference the dynamic Prometheus datasource, resolving issues with improper binding. [#2287](https://github.com/openfga/openfga/issues/2287)
 - Handle case where iterator is stopped more than once for `current_iterator_cache_count`. [#2409](https://github.com/openfga/openfga/pull/2409)
+- Fix deadlock when number of SQL datastore connections is less than Resolve Max Breadth. [#2425](https://github.com/openfga/openfga/pull/2425)
+- Improved `panic()` handling from `go` routines: [#2379](https://github.com/openfga/openfga/pull/2379), [#2385](https://github.com/openfga/openfga/pull/2385), [#2405](https://github.com/openfga/openfga/pull/2405), [#2428](https://github.com/openfga/openfga/pull/2428)
 
 ## [1.8.11] - 2025-04-29
 [Full changelog](https://github.com/openfga/openfga/compare/v1.8.10...v1.8.11)
@@ -1282,7 +1295,8 @@ Re-release of `v0.3.5` because the go module proxy cached a prior commit of the 
 - Memory storage adapter implementation
 - Early support for preshared key or OIDC authentication methods
 
-[Unreleased]: https://github.com/openfga/openfga/compare/v1.8.11...HEAD
+[Unreleased]: https://github.com/openfga/openfga/compare/v1.8.12...HEAD
+[1.8.12]: https://github.com/openfga/openfga/compare/v1.8.11...v1.8.12
 [1.8.11]: https://github.com/openfga/openfga/compare/v1.8.10...v1.8.11
 [1.8.10]: https://github.com/openfga/openfga/compare/v1.8.9...v1.8.10
 [1.8.9]: https://github.com/openfga/openfga/compare/v1.8.8...v1.8.9